+++ HowTheyWrite_android/app/src/main/java/com/kaiserpudding/howtheywrite/characterList/CharacterListFragment.kt
@@ -120,7 +120,7 @@ class CharacterListFragment
         }
     }
 
-    override fun onListInteraction(itemId: Int) {
+    override fun onListInteraction(itemId: Long) {
         listener?.onCharacterListItemInteraction(itemId, 0)
     }
 }
