+++ HowTheyWrite_android/app/src/main/java/com/kaiserpudding/howtheywrite/lessonList/LessonListViewModel.kt
@@ -41,7 +41,7 @@ class LessonListViewModel(application: Application) : AndroidViewModel(applicati
         }
     }
 
-    fun deleteLessons(lessonIds: IntArray) {
+    fun deleteLessons(lessonIds: LongArray) {
         viewModelScope.launch {
             lessonRepository.deleteLessons(lessonIds)
         }
