+++ HowTheyWrite_android/app/src/main/java/com/kaiserpudding/howtheywrite/quiz/QuizViewModelFactory.kt
@@ -13,23 +13,23 @@ import androidx.lifecycle.ViewModelProvider
 class QuizViewModelFactory : ViewModelProvider.Factory {
 
     private val application: Application
-    private val lessonId: Int
-    private val characterIds: IntArray?
+    private val lessonId: Long
+    private val characterIds: LongArray?
 
-    constructor(application: Application, lessonId: Int) {
+    constructor(application: Application, lessonId: Long) {
         this.application = application
         this.lessonId = lessonId
         this.characterIds = null
     }
 
-    constructor(application: Application, characterIds: IntArray) {
+    constructor(application: Application, characterIds: LongArray) {
         this.application = application
         this.lessonId = -1
         this.characterIds = characterIds
     }
 
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return if (lessonId != -1) QuizViewModel(application, lessonId) as T
+        return if (lessonId != -1L) QuizViewModel(application, lessonId) as T
         else QuizViewModel(application, characterIds!!) as T
     }
 }
